Since the tests work in Eclipse, but not in `gradlew test`, I wanted to see if the plugin worked when it was actually used.
I added the 'maven-plugin' so that I could deploy to mavenLocal(), and the first problem was that the eclipse jars couldn't be resolved.  People will be wary of a plugin that needs them to change their buildscript classpath to some weird ivy thing, so I decided to go the fat jar route.

Now users of the plugin don't need to worry about the ivy nonsense, and it actually works!!  But the groovy tests still fail.  This makes me suspicious of the groovy test runner.

Next-up I'm going to convert the thing entirely to Java to see if that fixes the local tests.