Fixed to handle the case of a file with a long diff.
Before this commit, if there was one file with a very large diff (e.g. a line ending problem), then the error message would just be "0 other files".  While fixing this, it became easier to refactor a lot of the logic in FormatTask.formatViolationsFor() and put it into DiffMessageFormatter, which has absorbed DiffUtils.

See DiffMessageFormatterTest for an overview of what the error messages look like for various cases.