Added some focus to `assertSpotlessCheckSucceeds`.
SpotlessCheck isn't supposed to change any files.  It's good that the tests confirm that this is true, but it adds a lot of work for the programmer to specify the contents manually.  If we wanted to test this facet, a better way would be to read all the files in the build automatically, then run `spotlessCheck`, then make sure none of them changed.  No reason to make the programmer write this manually.

This commit replaces `spotlessCheckSucceeds` with two more focused methods: `checkIsUpToDate(boolean upToDate)` and `checkRunsThenUpToDate()` which just calls:

```
checkIsUpToDate(true);
checkIsUpToDate(false);
```

In the next commit, we'll show the advantage of this by testing up-to-date checking.