Improve Windows OS check in PaddedCellTaskTest
Before this commit, the "os.name" system property was parsed manually
to figure out if the test was running in Windows or not. However, I
seem to remember reading somewhere a long time ago that parsing
"os.name" is error prone and easy to mess up. So I'm pulling in a
dependency on Apache Commons Lang so that we can use
SystemUtils.IS_OS_WINDOWS, which means (1) we're less likely to have
flaky or misbehaving tests, and (2) if our tests do flake or misbehave,
and we determine that the bug is in Commons Lang, then we can be good
Samaritans and file a bug report for everyone's benefit.