Revert "Explicitly made Provisioner a functional interface."
This reverts commit 617cc0e14832373dbf9fe9d7b59cd289663a98fd.

Implementing a Provisioner is pretty complicated.  It does happen to be a one-method interface at the moment, and likely will remain one.  But if someone needs to pass a Provisioner as an argument, marking it as @FunctionalInterface indicates to them that they can do so with a lambda.  And that's not actually true, they'll almost definitely need to pass GradleProvisioner.createFromProject, or whatever the host build system's equivalent would be.