Removed LicenseHeaderStep#name and replaced with code duplication.
It doesn't seem wise to expose something in spotless-lib which is only
used in one place in plugin-gradle and which may be rendered useless if
plugin-gradle's implementation details change.

This is especially considering that spotless-lib may be used by people
other than ourselves, and so we will want to ensure we can evolve
spotless-lib in a semver-compatible way that doesn't inconvenience
other users. Therefore I believe that replacing LicenseHeaderStep#name
with the code duplication I introduced in this commit is a worthwhile
trade-off.