Detect all Java source by default
Fixes #59, by using the getAllJava() method on SourceSet instead
of getJava(). The main application where this is useful is projects
that do joint Java/Groovy compilation, which sometimes requires putting
Java source into the Groovy directory. getAllJava() correctly picks up
all of these source files.

The test is a modified copy of the GoogleJavaFormatTest. The main oddity
in this test is that a side-effect of the GradleIntegrationTest.write()
methods is that a newline is always present at the end of the file. This
means that you can't round trip a file through write()/read() and
expect it to be the same. To avoid modifying that behavior, since I'm
not sure how much other tests depend on it, I just re-read the groovy
source after writing to ensure I could verify it wasn't changed by
spotlessApply.