Fixed license header tests
They were previously not executed because of missing `@Test`
annotation. When enabled, they failed because maven was not able to
find `LicenseHeader` implementation. Looks like Maven is only able
to locate classes for injection when they live in the same package
or in subpackages of the current package. For example step `Java`
lives in package `com.diffplug.spotless.maven.java` and to be able
to inject implementation of `licenseHeader` step it should also live
in `com.diffplug.spotless.maven.java` or any of it's subpackages,
like `com.diffplug.spotless.maven.java.hello`.

This commit makes generic license header an abstract class and adds
implementor for Java in the correct package. Such implementations
are now also responsible for default delimiter. This feels cleaner
than putting delimiter in `FormatterFactory` implementations.