Removed need for "<steps>" config element
And need to restructure packages. Turns out Maven (underlying Eclipse
sisu container) requires injected classes be in the same package as
the class that's being configured. That's only in case of direct field
injection. With setter/adder is class can be in any package because it's
`Class` will be determined based on the type of setter/adder parameter.
It is also guaraneteed that setters and adders will be invoked in the
order of corresponding xml elements. This removes the need for `<steps>`
and makes it possible to place classes in nice packages. One drawback is
that explicit adder methods are required. At the same time they provide
an ability to explicitly define which steps are supported for the given
formatter. This makes error messages a bit prettier when for example
"googleJavaFormat" is configured for Scala.

Eclipse sisu doc for adder/setter:
https://www.eclipse.org/sisu/docs/api/org.eclipse.sisu.plexus/reference/org/eclipse/sisu/plexus/CompositeBeanHelper.html