Refactor LicenseHeaderStepTest for custom licenses
Refactor `LicenseHeaderStepTest` to allow to create license and class
files with custom license contents (defined in the class itself), to
make it easier to test new license contents without adding more files.

The license and class file templates (`LicenseHeaderWithPlaceholder` and
`FileWithLicenseHeaderAndPlaceholder`, respectively) contain a token
which is replaced with the custom license contents when running the
tests.
The year related license is now defined in `LicenseHeaderStepTest` and
the tests are changed to use it. Also, two more cases were added when
testing the year token, to improve the coverage of the condition that
checks if the year is properly defined in the license.

Add a method to `ResourceHarness` that allows to create a test file with
its contents processed by a function, to support the refactoring.