Support multiple formats in Maven plugin
Format is a configuration element that makes it possible to apply
generic formatting steps (EndWithNewline, LicenseHeader, etc.) to
a set of files defined by custom includes and excludes. Example of
format configuration:

```
<configuration>
  <format>
    <includes>
      <include>src/**/txt/**/*.txt</include>
    </includes>
    <excludes>
      <exclude>src/**/txt/**/*Wrong.txt</exclude>
    </excludes>
    <endWithNewline/>
    <trimTrailingWhitespace/>
    <licenseHeader>
      <content>// License Text</content>
      <delimiter>start</delimiter>
    </licenseHeader>
  </format>
</configuration>
```

It is currently possible to only define a single `<format>` element in
plugin configuration.

This commit makes it possible to define multiple formats. Format
configuration structure remains the same but all formats should be
enclosed in `<formats>...</formats>`. Example:

```
<configuration>
  <formats>
    <format>...format 1 includes, excludes and steps</format>
    <format>...format 2 includes, excludes and steps</format>
    <format>...format 3 includes, excludes and steps</format>
  </formats>
</configuration>
```