Let spotless plugin not break clean task with the configuration cache
Using the project at execution time is not allowed when Gradle's
configuration cache is enabled.

Before this commit, the spotless plugin registered a doLast task action
on the clean task that accessed the projet at execution time. The effect
of this is to break using the clean task on all builds that have the
spotless plugin applied when the configuration cache is enabled.

This commits change the way the spotless plugin configures the clean
task to not access a project in the added doLast task action by
computing the spotless cache key upfront at configuration time. The
effect is that builds with the spotless plugin applied can now run the
clean task with the configuration cache enabled.