Add a Gherkin formatter
As part of #907, we want to be able to format our Gherkin source files
automagically.

To do this, we can utilise the upstream gherkin-formatter project to do
the heavy lifting, and add a few test cases within this project to make
sure that functionality looks correct.

We're calling this a `simple` formatter as it doesn't allow much to be
configured other than the indentation size in spaces.