Up-to-date index for Maven plugin
POC of a component that allows the Maven plugin to skip already formatted and
unchanged files. The code is not production-ready and probably does not work.
It is only to illustrate the approach and gather feedback.

A new `UpToDateChecker` maintains an index file in the build directory
(`<project>/target`). The index file has the following structure:

```
<source file 1> <last modified time 1>
<source file 2> <last modified time 2>
...
<source file n> <last modified time n>
```

`UpToDateChecker` loads or creates an index file at creation time into an
in-memory map.

Method `#isUpToDate(File)` checks the file's current last modified time
against the last modified time from the in-memory map.

Method `#setUpToDate(File)` updates or adds the file and its last modified
time to the in-memory map.

Method `#close()` writes the complete in-memory map into the index file.

Spotless apply goal uses `UpToDateChecker` to determine if a file needs to be
formatted. It also notifies the checker if the file has was formatted.