Make Maven file index care about formatters
Include the list of serialized formatters in PluginFingerprint. It makes the
file index sensitive to changes in configuration files referenced by the POM
(e.g. `eclipseformat.xml` and `.gitattributes`).

To achieve this, Maven plugin has to instantiate all formatters upfront.
It builds a map of formatters and files that need formatting. The formatters
are then used to calculate a PluginFingerprint.