Add flexmark-java step
Introduces a new step to natively format markdown. Previously markdown
was only available to spotless through the use of prettier which
requires nodejs. By natively supporting markdown, java users of spotless
have one less dependency to worry about.

This uses https://github.com/vsch/flexmark-java, a fork of
commonmark-java that made it possible to parse many flavors of markdown
into an AST and render it in some way. Among those renderers, there is
the option to format to markdown.

This renderer (or formatter) supports many formatting options. I've
selected a few reasonable ones to get started.

A test has been added which is inspired by the example from
https://github.com/vsch/flexmark-java/wiki/Markdown-Formatter. I've
extended it with some interesting things the formatter seems to take
care of.

During development, I had some difficulties dealing with idempotency of
the step. This is because the parser and formatter can be configured to
use different flavors individually. Care should be taken to apply the
same configurations to both the parser and formatter to keep it
idempotent.