Handle overflow when converting FileTime to Instant
This conversion is needed because FileIndex stores human-readable timestamps in
the index file. There's a way to convert FileTime to String but no way to
convert String to FileTime when reading an index from a file. Thus, we convert
to Instant that has well-defined `#toString()` and `#parse()` methods.

Conversion from FileTime to Instant can be lossy because FileTime can represent
timestamps further in the past/future than Instant. Such timestamps are saturated
to Instant.MIN and Instant.MAX. It is not safe to record such timestamps in the
index file because they are lossy.

IndexBasedChecker will not record overflown Instants in the index. It will also
log a warning for every file with a small/large last modified time that overflows
an Instant.