Fix something from 15->17 in #1094
spotlessCheck fails on Java 17 unless we add params to gradle.properties, but we removed that in e578a9d2c9a7117d012a8c6e31c5f27f4d3a1600 because it breaks Java 8. The easy fix is to not run spotlessCheck to `test_nomaven`.

This was missed in #1094 because buildcache covered it up.