Add support for git's `core.autocrlf`
Among others, Spotless uses .gitattributes to determine the line ending
of files. In the absence of this file, Spotless falls back to `core.eol`
from git config.

Before this commit, Spotless ignored the `core.autocrlf` property. Git
behaves in such a way that it ignores `core.eol` if `core.autocrlf` is
set to `true` or `input`. This commit aligns the behavior of Spotless to
favor `core.autocrlf` over `core.eol`.

Fixes #540

Signed-off-by: Roland Weisleder <roland.weisleder@googlemail.com>