Migrate to SLF4J for logging
As noted in #1116, it'd be ideal if we moved to SLF4J as our logging
interface.

We can do this pretty easily, making sure we remove our `package-list`
reference, as well as making sure the SLF4J API is only used for
compilation.

We also need to map the error logging levels that most closely mirror
what `java.util.logging` uses.

Closes #1116.