Improve selectivity in Maven plugin fingerprint
`PluginFingerprint` captures a particular configuration of the Spotless Maven
plugin. It is used to invalidate the up-to-date index when plugin's configuration
changes.

Invalidations of the index file seem to be too frequent because `PluginFingerprint`
includes the effective (as in Effective POM) configuration of the Spotless plugin.
For example, it includes dependencies declared outside of the `<plugin></plugin>`
XML element. So addition of an unrelated project dependency will invalidate the
index file.

Spotless Maven plugin manages its own dependencies using `ArtifactResolver`. There's
no need to include the effective `<plugin>` object into `PluginFingerprint`. This
commit simplifies the fingerprint to include only the plugin's version and serialized
configuration of all formatters. It should be backward-compatible because fingerprint
format change will simply trigger an invalidation of the index file.