Enable up-to-date checking in Maven plugin by default
Up-to-date checking has been supported for a long time. We haven't received any
negative feedback about it. Neo4j is an example project that has up-to-date
checking enabled to improve performance and haven't had problems with it:
https://github.com/neo4j/neo4j/blob/6ebc409772cea13354adb6a11aab4ed38da9d53e/pom.xml#L745-L747.

It should be safe to enable up-to-date checking for all users of the Spotless Maven plugin.

This commit also fixes the problem where plugin execution would fail on a parent of a multimodule
Maven project with up-to-date checking enabled. A parent project can configure Spotless as:

```xml
<build>
  <pluginManagement>
    <plugins>
      <plugin>
        <groupId>com.diffplug.spotless</groupId>
        <artifactId>spotless-maven-plugin</artifactId>
        ...
      </plugin>
    </plugins>
  </pluginManagement>
</build>
```

It allows child projects to inherit plugin's version and configuration but the
plugin is not enables by default.

PluginFingerprint did not handle such configuration correctly. It could only handle a simple
configuration like:

```xml
<build>
  <plugins>
    <plugin>
      <groupId>com.diffplug.spotless</groupId>
      <artifactId>spotless-maven-plugin</artifactId>
      ...
    </plugin>
  </plugins>
</build>
```

And failed with "Spotless plugin absent from the project" error for a multimodule
parent project with Spotless in `<pluginManagement>`.