Check if ktlint_code_style is set in .editorconfig before overriding it
ktlint_code_style gets unconditionally overridden to its default value
(intellij_idea) when the editorConfigOverride map is non-empty but does
not define it. As a result, it gets overridden even if it's actually
defined in the .editorconfig file.

This change checks if ktlint_code_style is already defined in
.editorconfig before overriding it to its default value.

Fixes https://github.com/diffplug/spotless/issues/2142.