Use Gradle user home directory for downloading biome, fixes #2187
Use the Gradle user home directory by default for the download directory for the biome executable.
Previously, the plugin tried to use Maven's home directory, which is not always accessible by a Gradle plugin.
According to the [Gradle docs](https://docs.gradle.org/current/kotlin-dsl/gradle/org.gradle.api.invocation/-gradle/get-gradle-user-home-dir.html),
the user home directory

> is used to cache downloaded resources, compiled build scripts and so on.

From that description, it seems to me to be the proper location for the biome executable.