Improve diff calculation after failed RDF content verification
Comparing graph nodes using their `equals()` method is used under the hood by the isomorphicity check and therefore has to be used in the diff calcluation as well (otherwise the latter might not find any differences).

Special handling of blank nodes has been omitted as it is hard to find which ones are different and which ones are not, so we just print them all for now.