chore: make sure to calculate different hashes if serve-files change
this was a hidden update issue: if we change the serve-script content
between two releases but the user keeps the package.json the same, the
md5 hash used for the node-modules-dir stayed the same, resulting in us
not using the latest version of the serve script but keep using the old
one. That was especially harmful in the case where we changed the api of
the serve-script to accept a uuid to allow multiple instances. The old
script just ignored that resulting in a server-process that was never
found due to wrong port-file-names written/waited for.